requirement.txt
disnake
python-dotenv
psutil
uptime
feedparser
requests
APScheduler

